tools/autoposter/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
Pillow>=10.4.0
feedparser>=6.0.11
google-api-python-client>=2.137.0
google-auth-oauthlib>=1.2.1
google-auth-httplib2>=0.2.0
